Reorganize to move 3D specific functionality into java3d sub-package. Add some javadoc.
