+++ core/src/visad/java3d/DisplayImplJ3D.java
@@ -72,6 +72,8 @@ import java.awt.*;
 
 import javax.media.j3d.*;
 import com.sun.j3d.utils.applet.MainFrame;
+import java.util.Iterator;
+import java.util.Vector;
 // import com.sun.j3d.utils.applet.AppletFrame;
 
 /**
@@ -757,6 +759,76 @@ public class DisplayImplJ3D extends DisplayImpl {
     projection = null;
     mode = null;
   }
+  
+  float getOffsetDepthMinimum(float depthOffsetMax) {
+    Vector rendVec = getRendererVector();
+    Iterator<DataRenderer> iter = rendVec.iterator();
+    float offsetMin = depthOffsetMax;
+    while (iter.hasNext()) {
+      DataRenderer rend = iter.next();
+        if (rend.hasPolygonOffset()) {
+          if (rend.getPolygonOffset() < offsetMin) {
+            offsetMin = rend.getPolygonOffset();  
+          }
+        }
+    }
+    return offsetMin;
+  }
+    
+   int getNumRenderersWithZoffset() {
+     Vector rendVec = getRendererVector();
+     Iterator<DataRenderer> iter = rendVec.iterator();
+     int num = 0;
+     while (iter.hasNext()) {
+       DataRenderer rend = iter.next();
+       if (rend.hasPolygonOffset()) {
+         num++;
+       }
+     }
+     return num;
+   }
+   
+   /**
+    * Sets the depth buffer offset when autoDepthOffset is enabled for this display.
+    * @param renderer
+    * @param mode 
+    */
+   public void setDepthBufferOffset(DataRenderer renderer) {
+     if (mode.getAutoDepthOffsetEnable()) {
+       float depthOffsetInc = mode.getDepthOffsetIncrement();
+       int numLayers = mode.getNumRenderersWithDepthOffset();
+       float maxDepthOffset = numLayers*(-depthOffsetInc);
+       
+       if (!renderer.hasPolygonOffset()) {
+         int cnt = getNumRenderersWithZoffset();
+         if (cnt < numLayers) {
+           renderer.setPolygonOffset(getOffsetDepthMinimum(maxDepthOffset) + depthOffsetInc);
+           renderer.setPolygonOffsetFactor(0f);
+           renderer.setHasPolygonOffset(true);  
+         }
+         else {
+           renderer.setPolygonOffset(0f);  
+           renderer.setPolygonOffsetFactor(0f);
+           renderer.setHasPolygonOffset(false);
+         }
+       }
+       mode.setPolygonOffset(renderer.getPolygonOffset(), false);
+       mode.setPolygonOffsetFactor(renderer.getPolygonOffsetFactor(), false);
+     }
+   }
+   
+   public void resetDepthBufferOffsets() {
+     Vector rendVec = getRendererVector();
+     Iterator<DataRenderer> iter = rendVec.iterator();
+     while (iter.hasNext()) {
+       DataRenderer rend = iter.next();
+       if (rend.hasPolygonOffset()) {
+         rend.setHasPolygonOffset(false);
+         rend.setPolygonOffset(0f);
+         rend.setPolygonOffsetFactor(0f);
+       }
+     }
+   }
   
 }
